Remove `@finos/perspective-jupyterlab` TypeScript, replace with JavaScript
