+++ packages/perspective-jupyterlab/package.json
@@ -47,13 +47,10 @@
     "devDependencies": {
         "@finos/perspective-test": "^0.10.3",
         "@jupyter-widgets/base-manager": "^1.0.0-alpha.0",
-        "@types/jest": "^23.3.9",
-        "@types/node": "^11.11.0",
         "identity-obj-proxy": "^3.0.0",
         "isomorphic-fetch": "^2.2.1",
         "jest-transform-css": "^2.0.0",
-        "source-map-support": "^0.5.9",
-        "ts-jest": "^25.1.0"
+        "source-map-support": "^0.5.9"
     },
     "jupyterlab": {
         "extension": true
