+++ packages/perspective-jupyterlab/src/config/plugin.config.js
@@ -13,7 +13,7 @@ const PerspectivePlugin = require("@finos/perspective-webpack-plugin");
 module.exports = {
     mode: process.env.PSP_NO_MINIFY || process.env.PSP_DEBUG ? "development" : process.env.NODE_ENV || "production",
     entry: {
-        index: "./src/ts/index.ts"
+        index: "./src/js/index.js",
     },
     resolve: {
         extensions: [".ts", ".js"],
@@ -51,12 +51,7 @@ module.exports = {
                     options: {}
                 }
             },
-            {
-                test: /\.ts$/,
-                exclude: /node_modules/,
-                loader: "ts-loader"
-            }
-        ]
+        ],
     },
     externalsPresets: {web: false, webAsync: true},
     output: {
