+++ packages/perspective-jupyterlab/src/js/client.js
@@ -7,19 +7,7 @@
  *
  */
 
-/* eslint-disable @typescript-eslint/no-explicit-any */
-
-import {DOMWidgetView} from "@jupyter-widgets/base";
 import {Client} from "@finos/perspective/dist/esm/api/client";
-
-/**
- * The schema for a message passed to and from `PerspectiveJupyterClient`.
- */
-export interface PerspectiveJupyterMessage {
-    id: number;
-    type: string;
-    data: any;
-}
 
 /**
  * `PerspectiveJupyterClient` acts as a message bus between the frontend and
@@ -31,15 +19,14 @@ export interface PerspectiveJupyterMessage {
  * `@finos/perspective/api`.
  */
 export class PerspectiveJupyterClient extends Client {
-    view: DOMWidgetView;
-
     /**
      * Create a new instance of the client.
      *
      * @param view {DOMWidgetView} the plugin view that can send messages to the
      * Python backend.
      */
-    constructor(view: DOMWidgetView) {
+
+    constructor(view) {
         super();
         this.view = view;
     }
@@ -52,10 +39,15 @@ export class PerspectiveJupyterClient extends Client {
      *
      * @param msg {any} the message to pass to the `PerspectiveManager`.
      */
-    send(msg: any): void {
+
+    send(msg) {
         // Handle calls to `update` with a binary by setting `binary_length`
         // to true, so the kernel knows to handle the arraybuffer properly.
-        if (msg.method === "update" && msg.args.length === 2 && msg.args[0] instanceof ArrayBuffer) {
+        if (
+            msg.method === "update" &&
+            msg.args.length === 2 &&
+            msg.args[0] instanceof ArrayBuffer
+        ) {
             const binary_msg = msg.args[0];
             const buffers = [binary_msg];
 
@@ -66,7 +58,6 @@ export class PerspectiveJupyterClient extends Client {
             // Remove the arraybuffer from the message args, so it can be
             // passed along in `buffers`.
             msg.args.shift();
-
             const serialized = JSON.stringify(msg);
 
             // Send the first update message over the Jupyter comm with
@@ -74,7 +65,7 @@ export class PerspectiveJupyterClient extends Client {
             this.view.send({
                 id: msg.id,
                 type: "cmd",
-                data: serialized
+                data: serialized,
             });
 
             // Send the second message with buffers.
@@ -84,7 +75,7 @@ export class PerspectiveJupyterClient extends Client {
             this.view.send({
                 id: msg.id,
                 type: "cmd",
-                data: JSON.stringify(msg)
+                data: JSON.stringify(msg),
             });
         }
     }
