+++ packages/perspective-jupyterlab/src/js/version.js
@@ -7,11 +7,5 @@
  *
  */
 
-/* defines */
-export const MIME_TYPE = "application/psp+json";
-
-export const PSP_CLASS = "PSPViewer";
-
-export const PSP_CONTAINER_CLASS = "PSPContainer";
-
-export const PSP_CONTAINER_CLASS_DARK = "PSPContainer-dark";
+const pkg_json = require("../../package.json");
+export const PERSPECTIVE_VERSION = pkg_json.version;
