--- packages/perspective-jupyterlab/src/ts/.eslintrc.json
@@ -1,43 +1,0 @@
-{
-    "parser": "@typescript-eslint/parser",
-    "extends":  [
-        "plugin:@typescript-eslint/recommended"
-    ],
-    "plugins": [
-        "prettier",
-        "@typescript-eslint/eslint-plugin"
-    ],
-    "env": {
-        "browser": true,
-        "commonjs": true,
-        "es6": true,
-        "node": true,
-        "jasmine": true,
-        "jest": true
-    },
-    "parserOptions": {
-        "ecmaVersion": 2017,
-        "ecmaFeatures": {},
-        "sourceType": "module",
-        "experimentalObjectRestSpread": true
-    },
-    "rules": {
-        "prettier/prettier": ["error", {
-            "printWidth": 200,
-            "tabWidth": 4,
-            "bracketSpacing": false
-        }],
-
-        "max-len": ["error", 200],
-        "no-const-assign": "error",
-        "no-this-before-super": "error",
-        "no-undef": "error",
-        "no-unreachable": "error",
-        "no-unused-vars": "error",
-        "constructor-super": "error",
-        "valid-typeof": "error",
-
-        "@typescript-eslint/camelcase": "off",
-        "@typescript-eslint/ban-ts-ignore": "off"
-    }
-}
