--- packages/perspective-jupyterlab/src/ts/arraybuffer.d.ts
@@ -1,12 +1,0 @@
-/******************************************************************************
- *
- * Copyright (c) 2018, the Perspective Authors.
- *
- * This file is part of the Perspective library, distributed under the terms of
- * the Apache License 2.0.  The full license can be found in the LICENSE file.
- *
- */
-
-declare module "@finos/perspective/dist/esm/api/client";
-declare module "@finos/perspective/dist/umd/psp.async.wasm";
-declare module "arraybuffer-loader!*";
