--- packages/perspective-jupyterlab/tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "compilerOptions": {
-    "noImplicitAny": true,
-    "noEmitOnError": true,
-    "noUnusedLocals": true,
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "allowJs": true,
-    "target": "es2016",
-    "outDir": "./build",
-    "lib": [
-      "es2016",
-      "ES2015.Promise",
-      "DOM"
-    ],
-    "types":[
-      "node"
-    ]
-  },
-  "include": ["package.json", "src/ts/*"]
-}
